AGENT: chunk history entries with dividers and HistoryChunk component
Extract HistoryItem into its own HistoryChunk component file and replace the
flat space-y-6 list with explicit horizontal rule dividers (my-10 md:my-12)
between entries. Within each chunk, the question Card and AnswerCard keep
tighter space-y-3 grouping, making each Q+A read as a single visual unit
with clear separation from the next.
